Reintroduce a shift buffer
I re-introduce a shift buffer that drops last trace in order to
make space to a new trace. However this will happen only if there is
no extra space left to extend the buffer. In fact, if garbage
collector is doing a proper work, this shift of elements should almost
never take affect.